requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
Pillow==10.1.0
numpy==1.24.3
watchdog==3.0.0
tqdm==4.66.4
ftfy==6.2.0
regex==2024.5.15
